requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
colorama==0.4.4
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-basic-stats==0.2.0
django-heroku==0.3.1
django-import-export==2.4.0
et-xmlfile==1.0.1
filelock==3.0.12
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
pipenv==2020.8.13
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
virtualenv==20.0.35
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
